Refactored http classes before SSE client implementation
