+++ Split.xcodeproj/xcshareddata/xcschemes/Split.xcscheme
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "3B6DEE5920EA6A4E0067435E"
-               BuildableName = "Split.framework"
-               BlueprintName = "Split"
-               ReferencedContainer = "container:Split.xcodeproj">
+               BlueprintIdentifier = "607FACCF1AFB9204008FA782"
+               BuildableName = "Split_Example.app"
+               BlueprintName = "Split_Example"
+               ReferencedContainer = "container:../sdk-tools/ios/debug/ios-dev-example/Split.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -73,6 +73,15 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
+      <BuildableProductRunnable>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "607FACCF1AFB9204008FA782"
+            BuildableName = "Split_Example.app"
+            BlueprintName = "Split_Example"
+            ReferencedContainer = "container:../sdk-tools/ios/debug/ios-dev-example/Split.xcodeproj">
+         </BuildableReference>
+      </BuildableProductRunnable>
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
@@ -82,6 +91,10 @@
             ReferencedContainer = "container:Split.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
+      <LocationScenarioReference
+         identifier = "com.apple.dt.IDEFoundation.CurrentLocationScenarioIdentifier"
+         referenceType = "1">
+      </LocationScenarioReference>
    </LaunchAction>
    <ProfileAction
       buildConfiguration = "Release"
