+++ Split/Network/RestClient/RestClient+TrackEvents.swift
@@ -8,11 +8,11 @@
 
 import Foundation
 
-protocol RestClientTrackEvents: RestClientProtocol {
+protocol RestClientTrackEvents: RestClient {
     func sendTrackEvents(events: [EventDTO], completion: @escaping (DataResult<EmptyValue>) -> Void)
 }
 
-extension RestClient: RestClientTrackEvents {
+extension DefaultRestClient: RestClientTrackEvents {
   func sendTrackEvents(events: [EventDTO], completion: @escaping (DataResult<EmptyValue>) -> Void) {
     self.execute(target: EnvironmentTargetManager.sendTrackEvents(events: events), completion: completion)
     }
